refactor: move item core and base CSS to styles folder
