+++ packages/avatar-group/src/styles/vaadin-avatar-group-menu-item-base-styles.js
@@ -4,7 +4,7 @@
  * This program is available under Apache License Version 2.0, available at https://vaadin.com/license/
  */
 import { css } from 'lit';
-import { itemStyles } from '@vaadin/item/src/vaadin-item-base-styles.js';
+import { itemStyles } from '@vaadin/item/src/styles/vaadin-item-base-styles.js';
 
 const menuItemStyles = css`
   @layer base {
