+++ packages/avatar-group/src/styles/vaadin-avatar-group-menu-item-core-styles.js
@@ -3,6 +3,6 @@
  * Copyright (c) 2020 - 2025 Vaadin Ltd.
  * This program is available under Apache License Version 2.0, available at https://vaadin.com/license/
  */
-import { itemStyles } from '@vaadin/item/src/vaadin-item-core-styles.js';
+import { itemStyles } from '@vaadin/item/src/styles/vaadin-item-core-styles.js';
 
 export const avatarGroupMenuItemStyles = itemStyles;
