+++ packages/item/src/styles/vaadin-item-base-styles.js
