+++ packages/item/src/styles/vaadin-item-core-styles.d.ts
