+++ packages/item/src/styles/vaadin-item-core-styles.js
