+++ packages/item/src/vaadin-item.js
@@ -9,7 +9,7 @@ import { DirMixin } from '@vaadin/component-base/src/dir-mixin.js';
 import { PolylitMixin } from '@vaadin/component-base/src/polylit-mixin.js';
 import { CSSInjectionMixin } from '@vaadin/vaadin-themable-mixin/css-injection-mixin.js';
 import { ThemableMixin } from '@vaadin/vaadin-themable-mixin/vaadin-themable-mixin.js';
-import { itemStyles } from './vaadin-item-core-styles.js';
+import { itemStyles } from './styles/vaadin-item-core-styles.js';
 import { ItemMixin } from './vaadin-item-mixin.js';
 
 /**
